first pass at #48 ("Need to be able to see all 900+ eggs")
My initial approach is to a) move the fetching of feeds out of the root
route and into an JSON AJAX API one and b) add a caching layer. A
slightly better approach might be to load 100 eggs at a time through
javascript (this approach loads all 900+ at a time).